+++ packages/features/tailwind/index.ts
@@ -1,16 +1,17 @@
-import { RemedyFeatureFactory, createHtmlElement, info } from "@niarada/remedy";
+import { RemedyFeatureFactory, info } from "@niarada/remedy";
 import typography from "@tailwindcss/typography";
 import autoprefixer from "autoprefixer";
-import { existsSync, writeFileSync } from "node:fs";
+import { existsSync, readFileSync, writeFileSync } from "node:fs";
 import postcss from "postcss";
 import tailwind, { Config as TailwindConfig } from "tailwindcss";
+import nesting from "tailwindcss/nesting";
 
 const defaultTailwindConfig = `export default {
     content: ["./**/*.part"],
 }`;
 
 export default function (): RemedyFeatureFactory {
-    return () => {
+    return (config) => {
         if (!existsSync("tailwind.config.js") && !existsSync("tailwind.config.ts")) {
             info("tailwind", "creating 'tailwind.config.ts'");
             writeFileSync("tailwind.config.ts", defaultTailwindConfig);
@@ -19,7 +20,7 @@ export default function (): RemedyFeatureFactory {
         return {
             name: "tailwind",
             async intercede(context) {
-                if (context.url.pathname === "/_tailwind") {
+                if (context.url.pathname.endsWith(".css")) {
                     let tailwindConfig = {} as TailwindConfig;
                     if (existsSync("tailwind.config.ts")) {
                         tailwindConfig = (await import(`${process.cwd()}/tailwind.config.ts`)).default;
@@ -30,14 +31,11 @@ export default function (): RemedyFeatureFactory {
                     if (!tailwindConfig.plugins?.includes(typography)) {
                         tailwindConfig.plugins?.push(typography);
                     }
-                    const css = `
-                    @import "tailwindcss/base";
-                    @import "tailwindcss/components";
-                    @import "tailwindcss/utilities";
-                `;
-                    const processor = postcss([autoprefixer, tailwind(tailwindConfig)]);
+                    const css = readFileSync(`${config.public}${context.url.pathname}`, "utf-8");
+                    const processor = postcss([nesting, autoprefixer, tailwind(tailwindConfig)]);
                     const result = await processor.process(css, {
-                        from: "<builtin>",
+                        // from: "<builtin>",
+                        from: context.url.pathname,
                     });
                     return new Response(result.css, {
                         headers: {
@@ -46,17 +44,6 @@ export default function (): RemedyFeatureFactory {
                     });
                 }
                 return undefined;
-            },
-            transform(node) {
-                if (node.type === "element" && node.tag === "head") {
-                    node.children.push(
-                        createHtmlElement(node, "link", {
-                            rel: "stylesheet",
-                            href: "/_tailwind",
-                        }),
-                    );
-                }
-                return node;
             },
         };
     };